Admin buttons blocking tip button in ManageArtistButtons.tsx
I propose moving the admin buttons to the left.

When making a post on Mirlo, they block the "Tip" button. Also, they block the tip button when you preview it live. So for the longest time, I didn't even know there was a tip button on the page :)

This is a pull request that moves it to the left.